requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiomcache==0.7.0
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
black==21.12b0
blinker==1.4
charset-normalizer==2.0.10
click==8.0.3
databases==0.5.4
frozenlist==1.3.0
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
priority==2.0.0
pydantic==1.9.0
python-dotenv==0.19.2
quart==0.16.2
quart-auth==0.6.0
Quart-CORS==0.5.0
quart-csrf==0.2
SQLAlchemy==1.4.29
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
Werkzeug==2.0.2
wsproto==1.0.0
WTForms==3.0.1
yarl==1.7.2
